+++ test/Feature/TRegx/CleanRegex/PatternBuilder/builder/PatternBuilderTest.php
@@ -9,6 +9,21 @@ class PatternBuilderTest extends TestCase
     /**
      * @test
      */
+    public function shouldBuild_prepared()
+    {
+        // given
+        $pattern = PatternBuilder::builder()->prepare(['You/her, (are|is) ', ['real? (or are you not real?)'], ' (you|her)']);
+
+        // when
+        $pattern = $pattern->delimited();
+
+        // then
+        $this->assertEquals('#You/her, (are|is) real\?\ \(or\ are\ you\ not\ real\?\) (you|her)#', $pattern);
+    }
+
+    /**
+     * @test
+     */
     public function shouldBuild_bind()
     {
         // given